+++ modules/exceptions.py
@@ -1,0 +1,3 @@
+class RestartFlowException(Exception):
+    """Custom exception to restart from initial function when an error occurs."""
+    pass